back-end/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
Automat==22.10.0
Babel==2.14.0
bcrypt==4.2.0
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.3
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cssselect==1.2.0
cssutils==2.11.1
debugpy==1.8.1
decorator==5.1.1
dlib==19.24.6
executing==2.0.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==22.10.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
json5==0.9.25
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
lxml==5.2.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==2.1.1
opencv-python==4.10.0.84
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
pillow==10.4.0
premailer==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.1
psutil==5.9.8
pure-eval==0.2.2
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-http-client==3.3.7
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.13
pyzmq==26.0.2
qtconsole==5.5.1
QtPy==2.4.1
queuelib==1.6.2
requests==2.32.3
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Scrapy==2.11.1
Send2Trash==1.8.3
sendgrid==6.11.0
service-identity==24.1.0
setuptools==75.1.0
SQLAlchemy==2.0.35
stack-data==0.6.3
starkbank-ecdsa==2.2.0
terminado==0.18.1
tinycss2==1.3.0
tldextract==5.1.2
traitlets==5.14.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
w3lib==2.1.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.10
yagmail==0.15.293
zope.interface==6.3
